* Simplify HSpinBox a bit, hopefully fixing #1960906 Crash on tempo   change * Make MIDIInstrumentParameterPanel::slotToggleChangeListOnProgChange   an actual slot rather than just a function!
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@9490 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/base/Composition.cpp
@@ -1533,7 +1533,7 @@ Track* Composition::getTrackById(TrackId track) const
 	      << __FILE__ << ":" << __LINE__ << std::endl;
     std::cerr << "Available track ids are: " << std::endl;
     for (trackconstiterator i = m_tracks.begin(); i != m_tracks.end(); ++i) {
-	std::cerr << (*i).second->getId() << std::endl;
+	std::cerr << (int)(*i).second->getId() << std::endl;
     }
 
     return 0;
